Move supported currencies into Price namespace
