+++ tests/Price/SupportedCurrencies/NeedsSupportedCurrencies.php
@@ -1,16 +1,16 @@
 <?php
 declare(strict_types=1);
 
-namespace SwipeStripe\Tests\Price;
+namespace SwipeStripe\Tests\Price\SupportedCurrencies;
 
 use SilverStripe\Core\Config\Config;
 use SilverStripe\Core\Injector\Injector;
-use SwipeStripe\SupportedCurrencies\MultipleSupportedCurrencies;
-use SwipeStripe\SupportedCurrencies\SupportedCurrenciesInterface;
+use SwipeStripe\Price\SupportedCurrencies\MultipleSupportedCurrencies;
+use SwipeStripe\Price\SupportedCurrencies\SupportedCurrenciesInterface;
 
 /**
  * Trait NeedsSupportedCurrencies
- * @package SwipeStripe\Tests\Price
+ * @package SwipeStripe\Tests\Price\SupportedCurrencies
  */
 trait NeedsSupportedCurrencies
 {
