+++ tests/Price/SupportedCurrencies/SingleSupportedCurrencyTest.php
@@ -1,18 +1,18 @@
 <?php
 declare(strict_types=1);
 
-namespace SwipeStripe\Tests\SupportedCurrencies;
+namespace SwipeStripe\Tests\Price\SupportedCurrencies;
 
 use Money\Currency;
 use Money\Exception\UnknownCurrencyException;
 use Money\Money;
 use SilverStripe\Core\Config\Config;
 use SilverStripe\Dev\SapphireTest;
-use SwipeStripe\SupportedCurrencies\SingleSupportedCurrency;
+use SwipeStripe\Price\SupportedCurrencies\SingleSupportedCurrency;
 
 /**
  * Class SingleSupportedCurrencyTest
- * @package SwipeStripe\Tests\SupportedCurrencies
+ * @package SwipeStripe\Tests\Price\SupportedCurrencies
  */
 class SingleSupportedCurrencyTest extends SapphireTest
 {
